# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "描述",
        "title": "后台管理",
        "contact": {},
        "version": "1.0 版本"
    },
    "paths": {
        "/user": {
            "get": {
                "description": "产品列表",
                "tags": [
                    "产品管理"
                ],
                "summary": "产品列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "当前页码",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "分页数量",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "post": {
                "description": "添加产品",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "产品管理"
                ],
                "summary": "添加产品",
                "parameters": [
                    {
                        "description": "产品名称",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "产品价格",
                        "name": "price",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseJson"
                        },
                        "headers": {
                            "token": {
                                "type": "string",
                                "description": "token"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "response.ResponseJson": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}